+++ gymlib_package/gymlib/tests/integtest_replay.py
@@ -1,17 +1,17 @@
 import unittest
 
-from gymlib.workspace import DBGymWorkspace
-
-from benchmark.tpch.constants import DEFAULT_TPCH_SEED
-from env.replay import replay
-from env.tests.gymlib_integtest_util import GymlibIntegtestManager
-from env.tuning_artifacts import (
+from gymlib.replay import replay
+from gymlib.tests.gymlib_integtest_util import GymlibIntegtestManager
+from gymlib.tuning_artifacts import (
     DBMSConfigDelta,
     IndexesDelta,
     QueryKnobsDelta,
     SysKnobsDelta,
     TuningArtifactsWriter,
 )
+from gymlib.workspace import DBGymWorkspace
+
+from benchmark.tpch.constants import DEFAULT_TPCH_SEED
 
 
 class ReplayTests(unittest.TestCase):
